Add selection of rasterizer/SDF renderer using CLI args. Currently SDF renderer does nothing.
I'm keeping the SDF renderer in the same file, because there is a lot of common input handling and window setup between the renderers. I made a CLI args based selector for the renderers just for convenience. It would be annoying to recompile every time to swap the rendering mode.